refactor(pyo3): split types.rs unit tests into a sibling module
The preceding field-typing change pushed gen_bindings/types.rs past the 1,000-line cap.
Move its `mod tests` into `types/tests.rs` (mirroring `gen_stubs/enums/tests.rs`),
bringing the file back to 800 lines. Pure move — generated output is byte-identical.
